requirements.txt
async-timeout==4.0.3
numpy==1.26.4
pandas==2.2.1
progressbar2==4.4.1
python-dateutil==2.8.2
python-utils==3.8.2
pytz==2024.1
redis==5.0.1
six==1.16.0
typing_extensions==4.10.0
tzdata==2024.1
